Do not attempt to print an empty bailiwick
The protobuf schema allows the bailiwick to be empty (unlike a single 0x00
for the root zone, which is obviously valid).

For empty byte sequences, the protobuf decoder returns a NULL pointer and a
length set to 0.

Unfortunately, ldns' behavior is undefined when trying to decode a NULL pointer
as a name.